% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angenieux.R
\name{Angenieux}
\alias{Angenieux}
\title{R6 Class for visualising floundeR based datasets}
\description{
This class aims to provide aplanat-like visualisation abstraction for the
floundeR framework
}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{Angenieux}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{A method to dump out the stored data from an \code{Angenieux} object}

\item{\code{plot}}{A method to plot the information stored in the \code{Angenieux} object - the
plot itself is determined by the stored data and other parameters for
a consistent rendering and per-project control}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Angenieux$new()}}
\item \href{#method-.plot}{\code{Angenieux$.plot()}}
\item \href{#method-clone}{\code{Angenieux$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="bin_data">}\href{../../floundeR/html/FloundeR.html#method-bin_data}{\code{floundeR::FloundeR$bin_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print">}\href{../../floundeR/html/FloundeR.html#method-print}{\code{floundeR::FloundeR$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new Angenieux object. This
initialisation method performs other sanity checking of the defined
file(s) and creates the required data structures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$new(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{the datatype that is being passed to the object}

\item{\code{value}}{the data that is being passed to the object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Angenieux} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.plot"></a>}}
\if{latex}{\out{\hypertarget{method-.plot}{}}}
\subsection{Method \code{.plot()}}{
Prepare and present an Angenieux plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$.plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters passed on to downstream methods - please see
examples for further examples as to how Angenieux plots can be customised
using this approach.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
