% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_card.R
\name{bam_qc_report_card_thresholds}
\alias{bam_qc_report_card_thresholds}
\title{Default BAM report-card thresholds}
\usage{
bam_qc_report_card_thresholds()
}
\value{
A named list of warn/fail thresholds used by \code{bam_qc_report_card()}.
}
\description{
Default BAM report-card thresholds
}
