% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{fast5.get_exp_start_time}
\alias{fast5.get_exp_start_time}
\title{Extract the experiment start time identified within a FAST5 file}
\usage{
fast5.get_exp_start_time(path, atomic = TRUE)
}
\arguments{
\item{path}{The file.path to check}

\item{atomic}{whether a single entity (multi files only) should be considered
(TRUE)}
}
\value{
vector of experiment_start_time(s)
}
\description{
This method parses a FAST5 sequence file and extracts the information on
specified experiment start time - this can be used in palaeogenomics of such
ancient datasets
}
\examples{
multiFast5 <- system.file("extdata", "multi.fast5", package="floundeR")
fast5.get_exp_start_time(multiFast5)

}
