% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{fast5.get_flowcell_id}
\alias{fast5.get_flowcell_id}
\title{Extract the flowcell_id identified within a FAST5 file}
\usage{
fast5.get_flowcell_id(path, atomic = TRUE)
}
\arguments{
\item{path}{The file.path to check}

\item{atomic}{whether a single entity (multi files only) should be considered
(TRUE)}
}
\value{
vector of flowcell_id(s)
}
\description{
This method parses a FAST5 sequence file and extracts the information on
specified flowcell_id - this can be used to track flowcells used within a
laboratory
}
\examples{
multiFast5 <- system.file("extdata", "multi.fast5", package="floundeR")
fast5.get_flowcell_id(multiFast5)

}
