% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{fast5.get_flowcell_type}
\alias{fast5.get_flowcell_type}
\title{Extract the flowcell type identified within a FAST5 file}
\usage{
fast5.get_flowcell_type(path, atomic = TRUE)
}
\arguments{
\item{path}{The file.path to check}

\item{atomic}{whether a single entity (multi files only) should be considered
(TRUE)}
}
\value{
vector of flowcell_types(s)
}
\description{
This method parses a FAST5 sequence file and extracts the information on
the flowcell type - this is based on user entry at start of run.
}
\examples{
multiFast5 <- system.file("extdata", "multi.fast5", package="floundeR")
fast5.get_flowcell_type(multiFast5)

}
