% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{fast5.get_info}
\alias{fast5.get_info}
\title{Extract salient experimental information from a FAST5 file}
\usage{
fast5.get_info(path, atomic = FALSE)
}
\arguments{
\item{path}{The file.path to check}

\item{atomic}{whether a single entity (multi files only) should be considered
(FALSE)}
}
\value{
data.frame
}
\description{
The FAST5 sequence file contains information relating to the sequencing
platform, flowcell and library preparation kits used in a study. This method
pulls these information into a data.frame or factor.
}
\examples{
promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
fast5.get_info(promFast5)

}
\seealso{
\code{\link{file.isSingleFast5} \link{file.isFast5}}
}
