% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{fast5.get_platform}
\alias{fast5.get_platform}
\title{Extract the sequencing platform identified within a FAST5 file}
\usage{
fast5.get_platform(path, atomic = TRUE)
}
\arguments{
\item{path}{The file.path to check}

\item{atomic}{whether a single entity (multi files only) should be considered
(TRUE)}
}
\value{
vector of platform(s)
}
\description{
This method parses a FAST5 sequence file and extracts the information on
specified sequencing platform - this can be used to identify e.g. gridion or
promethion based information.
}
\examples{
promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
fast5.get_platform(promFast5)

}
