% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{fast5.get_read_count}
\alias{fast5.get_read_count}
\title{Extract the number of reads identified within a FAST5 file}
\usage{
fast5.get_read_count(path)
}
\arguments{
\item{path}{The file.path to check}
}
\value{
integer count of reads
}
\description{
This method parses a FAST5 sequence file and extracts the information on
the read count - this is a simple count only
}
\examples{
multiFast5 <- system.file("extdata", "multi.fast5", package="floundeR")
fast5.get_read_count(multiFast5)

}
