% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{Fast5}
\alias{Fast5}
\title{R6 Class for loading and analysing nanopore FAST5 sequence files}
\description{
This class aims to simplify the handling and exploration of FAST5 files and
provides simple methods for accessing information that can be used to
identify the sequencing platform, flowcell used, sequencing kit and other
metainformation that includes whether they are single- or- multi format
files.
}
\examples{

## ------------------------------------------------
## Method `Fast5$is_single_fast5`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$is_single_fast5()

## ------------------------------------------------
## Method `Fast5$is_multi_fast5`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$is_multi_fast5()

## ------------------------------------------------
## Method `Fast5$get_platform`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_platform()

## ------------------------------------------------
## Method `Fast5$get_flowcell_id`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_flowcell_id()

## ------------------------------------------------
## Method `Fast5$get_exp_start_time`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_exp_start_time()

## ------------------------------------------------
## Method `Fast5$get_read_count`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_read_count()

## ------------------------------------------------
## Method `Fast5$get_flowcell_type`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_flowcell_type()

## ------------------------------------------------
## Method `Fast5$get_sequencing_kit`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_sequencing_kit()

## ------------------------------------------------
## Method `Fast5$get_info`
## ------------------------------------------------

promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_info(atomic=FALSE)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fast5_file}}{the file.path to the query FAST5 file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Fast5-new}{\code{Fast5$new()}}
\item \href{#method-Fast5-is_single_fast5}{\code{Fast5$is_single_fast5()}}
\item \href{#method-Fast5-is_multi_fast5}{\code{Fast5$is_multi_fast5()}}
\item \href{#method-Fast5-get_platform}{\code{Fast5$get_platform()}}
\item \href{#method-Fast5-get_flowcell_id}{\code{Fast5$get_flowcell_id()}}
\item \href{#method-Fast5-get_exp_start_time}{\code{Fast5$get_exp_start_time()}}
\item \href{#method-Fast5-get_read_count}{\code{Fast5$get_read_count()}}
\item \href{#method-Fast5-get_flowcell_type}{\code{Fast5$get_flowcell_type()}}
\item \href{#method-Fast5-get_sequencing_kit}{\code{Fast5$get_sequencing_kit()}}
\item \href{#method-Fast5-get_info}{\code{Fast5$get_info()}}
\item \href{#method-Fast5-print}{\code{Fast5$print()}}
\item \href{#method-Fast5-clone}{\code{Fast5$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-new"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-new}{}}}
\subsection{Method \code{new()}}{
Creates a new Fast5 object. This initialisation method performs other
sanity checking of the defined file to ensure that it is indeed a
file, a single file, is in HDF5 format and is either single- or
multi-FAST5 compliant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$new(fast5_file = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fast5_file}}{The source FAST5 file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Fast5} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-is_single_fast5"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-is_single_fast5}{}}}
\subsection{Method \code{is_single_fast5()}}{
Check if a provided file corresponds to a single entry FAST5 file

Simple method to check whether the provided file path corresponds to
a valid FAST5 file and if this file corresponds to a single FAST5
format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$is_single_fast5()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
boolean
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$is_single_fast5()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-is_multi_fast5"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-is_multi_fast5}{}}}
\subsection{Method \code{is_multi_fast5()}}{
Check if a provided file corresponds to a multiple entry FAST5 file

Simple method to check whether the provided file path corresponds to
a valid FAST5 file and if this file corresponds to a multi FAST5
format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$is_multi_fast5()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
boolean
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$is_multi_fast5()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-get_platform"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-get_platform}{}}}
\subsection{Method \code{get_platform()}}{
Extract the sequencing platform identified within a FAST5 file

This method parses a FAST5 sequence file and extracts the information
on specified sequencing platform - this can be used to identify e.g.
gridion or promethion based information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$get_platform()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of platform
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_platform()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-get_flowcell_id"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-get_flowcell_id}{}}}
\subsection{Method \code{get_flowcell_id()}}{
Extract the flowcell_id identified within a FAST5 file

This method parses a FAST5 sequence file and extracts the information
on specified flowcell_id - this can be used to track flowcells used
within a laboratory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$get_flowcell_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of flowcell_id
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_flowcell_id()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-get_exp_start_time"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-get_exp_start_time}{}}}
\subsection{Method \code{get_exp_start_time()}}{
Extract the experiment start time identified within a FAST5 file

This method parses a FAST5 sequence file and extracts the information
on specified experiment start time - this can be used in
palaeogenomics of such ancient datasets. Please note that the sampled
start time is transformed using \code{lubridate} for presentation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$get_exp_start_time()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of experiment_start_time
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_exp_start_time()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-get_read_count"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-get_read_count}{}}}
\subsection{Method \code{get_read_count()}}{
Extract the number of reads identified within a FAST5 file

This method parses a FAST5 sequence file and extracts the information
on the read count - this is a simple count only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$get_read_count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer count of reads
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_read_count()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-get_flowcell_type"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-get_flowcell_type}{}}}
\subsection{Method \code{get_flowcell_type()}}{
Extract the flowcell type identified within a FAST5 file

This method parses a FAST5 sequence file and extracts the information
on the flowcell type - this is based on user entry at start of run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$get_flowcell_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of flowcell_types
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_flowcell_type()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-get_sequencing_kit"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-get_sequencing_kit}{}}}
\subsection{Method \code{get_sequencing_kit()}}{
Extract the sequencing kit identified within a FAST5 file

This method parses a FAST5 sequence file and extracts the information
on the sequencing_kit used - this is based on user entry at start of
run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$get_sequencing_kit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of sequencing_kit
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_sequencing_kit()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-get_info}{}}}
\subsection{Method \code{get_info()}}{
Extract salient experimental information from a FAST5 file

The FAST5 sequence file contains information relating to the
sequencing platform, flowcell and library preparation kits used in a
study. This method pulls these information into a data.frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$get_info(atomic = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{atomic}}{whether a single entity (multi files only) should be
considered (FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{promFast5 <- system.file("extdata", "prom.fast5", package="floundeR")
promF5 <- Fast5$new(fast5_file=promFast5)
promF5$get_info(atomic=FALSE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-print"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-print}{}}}
\subsection{Method \code{print()}}{
print method to override the standard R6 printing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{stuff passed onwards}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fast5-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Fast5-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fast5$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
