% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{file.isFast5}
\alias{file.isFast5}
\title{Sanity check a file and ensure that it is a valid FAST5 file}
\usage{
file.isFast5(path)
}
\arguments{
\item{path}{The file.path to check}
}
\value{
boolean
}
\description{
Simple method to check whether the provided file path corresponds to a valid
FAST5 file (or not). This method operates only on atomic files.
}
\examples{
singleFast5 <- system.file("extdata", "single.fast5", package="floundeR")
file.isFast5(singleFast5)

}
