% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{file.isSingleFast5}
\alias{file.isSingleFast5}
\title{Check is a provided file corresponds to a single entry FAST5 file}
\usage{
file.isSingleFast5(path = NULL, f5content = NULL, silent = FALSE)
}
\arguments{
\item{path}{The file.path to check}

\item{f5content}{(optional) preparsed index from h5parser}

\item{silent}{should messages be hidden? (FALSE)}
}
\value{
boolean
}
\description{
Simple method to check whether the provided file path corresponds to a valid
FAST5 file and if this file corresponds to a single FAST5 format.
}
\examples{
singleFast5 <- system.file("extdata", "single.fast5", package="floundeR")
file.isSingleFast5(singleFast5)
multiFast5 <- system.file("extdata", "multi.fast5", package="floundeR")
file.isSingleFast5(multiFast5)

}
