% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fishy_fastq}
\alias{fishy_fastq}
\title{Consume a FASTQ file and collate basic characteristic observations}
\usage{
fishy_fastq(fastq)
}
\value{
DataFrame of FASTQ based sequence observations
}
\description{
This method is a fast C++ implementation of a FASTQ parser aimed to perform
a quick collation of summary_statistic information that may be used to make
decisions as to how FASTQ sequence collections are crafted.
}
\examples{
fastq <- system.file("extdata", "example.fastq.gz", package = "floundeR")
fishy_fastq(fastq)

}
