% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_check.R
\name{flnDr}
\alias{flnDr}
\title{extract a floundeR packaged file}
\usage{
flnDr(file)
}
\arguments{
\item{file}{a known filename (or its prefix)}
}
\value{
a file
}
\description{
A collection of files are distributed with the floundeR package; this
accessory method is aimed to abbreviate the technical code for the
identification of these files and to clean the vignettes and code examples.
}
