% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank_genome.R
\name{GenbankGenome}
\alias{GenbankGenome}
\title{R6 Class for loading and analysing Genbank whole genome files}
\description{
This class aims to implement a couple of trivial methods for wrangling whole
microbial genome information from Genbank files
}
\examples{

## ------------------------------------------------
## Method `GenbankGenome$new`
## ------------------------------------------------

TB_reference = flnDr("TB_H37Rv.gb.gz")
tb <- GenbankGenome$new(TB_reference)
}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{GenbankGenome}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenbankGenome-new}{\code{GenbankGenome$new()}}
\item \href{#method-GenbankGenome-list_cds}{\code{GenbankGenome$list_cds()}}
\item \href{#method-GenbankGenome-get_cds}{\code{GenbankGenome$get_cds()}}
\item \href{#method-GenbankGenome-focus_cds}{\code{GenbankGenome$focus_cds()}}
\item \href{#method-GenbankGenome-focus_range}{\code{GenbankGenome$focus_range()}}
\item \href{#method-GenbankGenome-as_tibble}{\code{GenbankGenome$as_tibble()}}
\item \href{#method-GenbankGenome-whole_genome_focus}{\code{GenbankGenome$whole_genome_focus()}}
\item \href{#method-GenbankGenome-focus_nucleotide}{\code{GenbankGenome$focus_nucleotide()}}
\item \href{#method-GenbankGenome-focus_codon}{\code{GenbankGenome$focus_codon()}}
\item \href{#method-GenbankGenome-get_nucleotide}{\code{GenbankGenome$get_nucleotide()}}
\item \href{#method-GenbankGenome-unfocus}{\code{GenbankGenome$unfocus()}}
\item \href{#method-GenbankGenome-is_focused}{\code{GenbankGenome$is_focused()}}
\item \href{#method-GenbankGenome-get_focus}{\code{GenbankGenome$get_focus()}}
\item \href{#method-GenbankGenome-clone}{\code{GenbankGenome$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="bin_data"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-bin_data'><code>floundeR::FloundeR$bin_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="num_scale"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-num_scale'><code>floundeR::FloundeR$num_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="nums_scale"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-nums_scale'><code>floundeR::FloundeR$nums_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-print'><code>floundeR::FloundeR$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-new}{}}}
\subsection{Method \code{new()}}{
Creates a new GenbankGenome object. This
initialisation method performs minimal sanity checking
of the defined file(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$new(gb_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gb_file}}{The source
sequencing_summary file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GenbankGenome} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{TB_reference = flnDr("TB_H37Rv.gb.gz")
tb <- GenbankGenome$new(TB_reference)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-list_cds"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-list_cds}{}}}
\subsection{Method \code{list_cds()}}{
exports the annotated genomic features as a GenomicRanges object.
This can be used for the review of content in a genomics context.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$list_cds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-get_cds"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-get_cds}{}}}
\subsection{Method \code{get_cds()}}{
Get the CDS information for one or more annotated features from the
genome of interest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$get_cds(feature_id = "fusA1")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-focus_cds"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-focus_cds}{}}}
\subsection{Method \code{focus_cds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$focus_cds(feature_id = "fusA1")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-focus_range"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-focus_range}{}}}
\subsection{Method \code{focus_range()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$focus_range(chromosome = NULL, start = NULL, end = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-whole_genome_focus"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-whole_genome_focus}{}}}
\subsection{Method \code{whole_genome_focus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$whole_genome_focus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-focus_nucleotide"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-focus_nucleotide}{}}}
\subsection{Method \code{focus_nucleotide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$focus_nucleotide(position)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-focus_codon"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-focus_codon}{}}}
\subsection{Method \code{focus_codon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$focus_codon(codon)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-get_nucleotide"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-get_nucleotide}{}}}
\subsection{Method \code{get_nucleotide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$get_nucleotide(position, strand)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-unfocus"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-unfocus}{}}}
\subsection{Method \code{unfocus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$unfocus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-is_focused"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-is_focused}{}}}
\subsection{Method \code{is_focused()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$is_focused()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-get_focus"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-get_focus}{}}}
\subsection{Method \code{get_focus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$get_focus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenbankGenome-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenbankGenome-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenbankGenome$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
