% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplex_set.R
\name{MultiplexSet}
\alias{MultiplexSet}
\title{R6 Class for loading, visualising and analysing barcode information}
\description{
R6 Class for loading, visualising and analysing barcode information

R6 Class for loading, visualising and analysing barcode information
}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{MultiplexSet}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{enumerate}}{prepares a simple \verb{2D Angenieux} enumeration of the provided dataset
for quick visualisation of the dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiplexSet-new}{\code{MultiplexSet$new()}}
\item \href{#method-MultiplexSet-as_tibble}{\code{MultiplexSet$as_tibble()}}
\item \href{#method-MultiplexSet-read_length_bins}{\code{MultiplexSet$read_length_bins()}}
\item \href{#method-MultiplexSet-quality_bins}{\code{MultiplexSet$quality_bins()}}
\item \href{#method-MultiplexSet-sequencingset}{\code{MultiplexSet$sequencingset()}}
\item \href{#method-MultiplexSet-temporalset}{\code{MultiplexSet$temporalset()}}
\item \href{#method-MultiplexSet-clone}{\code{MultiplexSet$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="bin_data"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-bin_data'><code>floundeR::FloundeR$bin_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="num_scale"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-num_scale'><code>floundeR::FloundeR$num_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="nums_scale"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-nums_scale'><code>floundeR::FloundeR$nums_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-print'><code>floundeR::FloundeR$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiplexSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiplexSet-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance of the R6 Class \code{MultiplexSet}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSet$new(seqsum = NA, barcoding_summary_file = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seqsum}}{a tibble of sequencing summary information}

\item{\code{barcoding_summary_file}}{is a file.path to the corresponding
barcoding_summary file that should be merged with the \code{seqsum}
content.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiplexSet-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-MultiplexSet-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Export the imported dataset(s) as a tibble

This object consumes a sequencing summary file (and optionally the
corresponding barcoding_summary file) and creates an object in
memory that can be explored, sliced and filtered. This method dumps
out the in-memory object for further exploration and development.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSet$as_tibble()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble representation of the starting dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiplexSet-read_length_bins"></a>}}
\if{latex}{\out{\hypertarget{method-MultiplexSet-read_length_bins}{}}}
\subsection{Method \code{read_length_bins()}}{
bin the sequences in \code{seqsum} content into bins of sequence length

The nanopore sequencing run is expected to return a collection of
sequences that vary in their length distributions; this variance is a
function of the sequencing library prepared, the starting DNA etc. This
method is used to bin reads into uniform bins & assess the distribution
of sequence lengths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSet$read_length_bins(
  qfilt = TRUE,
  normalised = TRUE,
  cumulative = FALSE,
  bins = 20,
  outliers = 0.025
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{qfilt}}{\itemize{
\item specifies how the quality information should be filtered
\item at the moment this only defines whether reporting is based on PASS
or FAIL reads; would make more sense to have filtered by PASS / ALL?
}}

\item{\code{normalised}}{\itemize{
\item should the sequence collection be reported to
normalise for the number of sequence bases sequenced or the number of
sequence reads - TRUE by default to normalise for sequenced bases.
}}

\item{\code{cumulative}}{defines whether cumulative sequence bases (reads) are
reported per bin (FALSE by default).}

\item{\code{bins}}{the number of sequence bins that should be prepared (20 by
default)}

\item{\code{outliers}}{defines the number of outliers (0.025 = 2.5\%) that are
excluded from the longest reads to prepare a richer distribution
visulation - the plots can be bothered by the long tail of mini-whales.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Angenieux 2D graph object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiplexSet-quality_bins"></a>}}
\if{latex}{\out{\hypertarget{method-MultiplexSet-quality_bins}{}}}
\subsection{Method \code{quality_bins()}}{
bin the sequences in \code{seqsum} content into bins of quality

The nanopore sequencing run is expected to return a collection of
sequences that vary in their quality distributions; this variance is a
function of the sequencing library prepared, the starting DNA etc. This
method is used to bin reads into uniform quality bins to assess the
overall quality of the run and to identify potential issues
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSet$quality_bins(bins = 20, outliers = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{the number of sequence bins that should be prepared (20 by
default)}

\item{\code{outliers}}{defines the number of outliers (0 = 0\%) that are
excluded from the reads - should probably be deprecated for simplicity?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Angenieux 2D graph object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiplexSet-sequencingset"></a>}}
\if{latex}{\out{\hypertarget{method-MultiplexSet-sequencingset}{}}}
\subsection{Method \code{sequencingset()}}{
Prepare a SequencingSet object from a given barcode

This method is used to subset the sequencing_summary information to
focus on a single barcode for more detailed analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSet$sequencingset(barcode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{barcode}}{the barcode that should be reported}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{SequencingSet} object containing information on barcode of
interest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiplexSet-temporalset"></a>}}
\if{latex}{\out{\hypertarget{method-MultiplexSet-temporalset}{}}}
\subsection{Method \code{temporalset()}}{
Prepare a TemporalSet object from a given barcode

This method is used to subset the SequencingSummary information to
focus on a single barcode for a more detailed analysis of content.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSet$temporalset(barcode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{barcode}}{the barcode that should be reported}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TemporalSet} object containing information on barcode of
interest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiplexSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiplexSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiplexSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
