% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_summary.R
\name{SequencingSummary}
\alias{SequencingSummary}
\title{R6 Class for loading and analysing nanopore sequencing_summary files}
\description{
This class aims to simplify the handling and exploration of
sequencing_summary files and provides simple methods for accessing
information that can be used to assess the performance of a run.
}
\examples{

## ------------------------------------------------
## Method `SequencingSummary$new`
## ------------------------------------------------

sequencing_summary <- flnDr("sequencing_summary.txt.bz2")
barcodes_summary <- flnDr("barcoding_summary.txt.bz2")
seqsum <- SequencingSummary$new(sequencing_summary, barcodes_summary)

## ------------------------------------------------
## Method `SequencingSummary$get_flowcell_platform`
## ------------------------------------------------

seqsum <- SequencingSummary$new(flnDr("sequencing_summary.txt.bz2"))
seqsum.get_flowcell_platform()

## ------------------------------------------------
## Method `SequencingSummary$get_fc_density_data`
## ------------------------------------------------

seqsum <- SequencingSummary$new(flnDr("sequencing_summary.txt.bz2"))
seqsum.get_fc_density_data()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sequencing_summary_file}}{the file.path
to the query FAST5 file}

\item{\code{barcoding_summary_file}}{the file.path
to the query FAST5 file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SequencingSummary$new()}}
\item \href{#method-get_flowcell_platform}{\code{SequencingSummary$get_flowcell_platform()}}
\item \href{#method-get_fc_density_data}{\code{SequencingSummary$get_fc_density_data()}}
\item \href{#method-clone}{\code{SequencingSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new SequencingSummary object. This
initialisation method performs other sanity checking
of the defined file(s) to ensure that it is indeed
parseable and creates the required data structures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$new(sequencing_summary_file, barcoding_summary_file = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sequencing_summary_file}}{The source
sequencing_summary file.}

\item{\code{barcoding_summary_file}}{The source
barcoding_summary_file file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Fast5} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{sequencing_summary <- flnDr("sequencing_summary.txt.bz2")
barcodes_summary <- flnDr("barcoding_summary.txt.bz2")
seqsum <- SequencingSummary$new(sequencing_summary, barcodes_summary)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_flowcell_platform"></a>}}
\if{latex}{\out{\hypertarget{method-get_flowcell_platform}{}}}
\subsection{Method \code{get_flowcell_platform()}}{
Have a guess at the most likely flowcell platform used

The sequencing summary file contains no information on the sequencing
device or flowcell used. For the preparation of channel density maps
it is worth considering which flowcell type is most likely to have
been used - this can be guessed on the number of channels described
within the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$get_flowcell_platform()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{seqsum <- SequencingSummary$new(flnDr("sequencing_summary.txt.bz2"))
seqsum.get_flowcell_platform()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fc_density_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_fc_density_data}{}}}
\subsection{Method \code{get_fc_density_data()}}{
produce channelMap for the predicted flowcell type for spatial plots

prepares a matrix of X, Y coordinates and
the corresponding readcount information for the type of flowcell
predicted by \code{get_flowcell_platform}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$get_fc_density_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
matrix with counts sorted by spatial rows and columns
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{seqsum <- SequencingSummary$new(flnDr("sequencing_summary.txt.bz2"))
seqsum.get_fc_density_data()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
