% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcell.R
\name{to_flowcell}
\alias{to_flowcell}
\title{Prepare a \code{flowcell} object from provided class}
\usage{
to_flowcell(r6_object)
}
\arguments{
\item{r6_object}{is the floundeR R6 object to extract flowcell from}
}
\value{
Flounder flowcell object
}
\description{
This method is intended to streamline the usage of floundeR in more coherent
workflows through the enabling of more magrittr-like piped commands. This
method will consume floundeR specified classes and return a flowcell object
where possible.
}
\examples{
SequencingSummary$new(flnDr("sequencing_summary.txt.bz2")) \%>\% to_flowcell()

}
