% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_set.R
\name{to_sequencing_set}
\alias{to_sequencing_set}
\title{Prepare a \code{SequencingSet} object from provided class}
\usage{
to_sequencing_set(r6_object)
}
\arguments{
\item{r6_object}{is the floundeR R6 object to extract SequencingSet from}
}
\value{
Flounder SequencingSet object
}
\description{
This method is intended to streamline the usage of floundeR in more coherent
workflows through the enabling of more magrittr-like piped commands. This
method will consume floundeR specified classes and return a SequencingSet
object where possible.
}
\examples{
SequencingSummary$new(flnDr("sequencing_summary.txt.bz2")) \%>\% 
    to_sequencing_set()

}
