% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angenieux.R
\name{Angenieux}
\alias{Angenieux}
\title{R6 Class for visualising floundeR based datasets}
\description{
This class aims to provide aplanat-like visualisation abstraction for the
floundeR framework
}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{Angenieux}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Angenieux$new()}}
\item \href{#method-XYDensity}{\code{Angenieux$XYDensity()}}
\item \href{#method-clone}{\code{Angenieux$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print">}\href{../../floundeR/html/FloundeR.html#method-print}{\code{floundeR::FloundeR$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$new(key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XYDensity"></a>}}
\if{latex}{\out{\hypertarget{method-XYDensity}{}}}
\subsection{Method \code{XYDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$XYDensity(matrix)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
