% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_blast.R
\name{Blast}
\alias{Blast}
\title{R6 Class for loading and analysing BLAST results in basic \code{Pairwise} format}
\description{
BLAST results are a fundamental unit of basic comparative genomics. This
R6 object has been implemented for the systematic exploration of BLAST
results within the scope of the Lodestar project.
}
\examples{

## ------------------------------------------------
## Method `Blast$new`
## ------------------------------------------------

blast_results <- flnDr("drosophila_uniref100.blastx.gz")
blast <- Blast$new(blast_file=blast_results)
}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{Blast}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blast_file}}{the file.path to the query BLAST results file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Blast$new()}}
\item \href{#method-count}{\code{Blast$count()}}
\item \href{#method-clone}{\code{Blast$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="bin_data">}\href{../../floundeR/html/FloundeR.html#method-bin_data}{\code{floundeR::FloundeR$bin_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="num_scale">}\href{../../floundeR/html/FloundeR.html#method-num_scale}{\code{floundeR::FloundeR$num_scale()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="nums_scale">}\href{../../floundeR/html/FloundeR.html#method-nums_scale}{\code{floundeR::FloundeR$nums_scale()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print">}\href{../../floundeR/html/FloundeR.html#method-print}{\code{floundeR::FloundeR$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new Blast object. This initialisation method performs other
sanity checking of the defined file(s) to ensure that it is indeed
parseable and creates the required data structures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Blast$new(blast_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blast_file}}{The source
sequencing_summary file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Blast} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{blast_results <- flnDr("drosophila_uniref100.blastx.gz")
blast <- Blast$new(blast_file=blast_results)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count"></a>}}
\if{latex}{\out{\hypertarget{method-count}{}}}
\subsection{Method \code{count()}}{
Return the number of BLAST results that are contained within the BLAST
file provided.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Blast$count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Blast$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
