% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floundeR.R
\name{FloundeR}
\alias{FloundeR}
\title{R6 Class for floundeR based analyses}
\description{
This is an overarching class which other flounder classes are expected to
inherit from. The \code{floundeR} object contains methods and variables that are
maintained for other objects - there is probably not a sensible usecase for
actually creating a \code{floundeR} instance on its own.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FloundeR$new()}}
\item \href{#method-print}{\code{FloundeR$print()}}
\item \href{#method-bin_data}{\code{FloundeR$bin_data()}}
\item \href{#method-clone}{\code{FloundeR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new FloundeR object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloundeR$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{FloundeR} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
this print method overrides the standard print function as included with
R6 objects - this is to better define what is contained within an object
and to provide better floundeR abstraction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloundeR$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional stuff passed on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing (at present) - output to stdout
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bin_data"></a>}}
\if{latex}{\out{\hypertarget{method-bin_data}{}}}
\subsection{Method \code{bin_data()}}{
general method for binning continuous data into sensible bins - used by
various methods within the broader floundeR framework.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloundeR$bin_data(data, bins = 20, outliers = 0.025, qmax = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\itemize{
\item the dataset (series) to be binned
}}

\item{\code{bins}}{\itemize{
\item the number of bins to prepare
}}

\item{\code{outliers}}{\itemize{
\item the fraction of longest outlying features that should
be excluded to simplify the plot.
}}

\item{\code{qmax}}{\itemize{
\item a manual upper limit to use
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
bin assignments for \code{data} elements
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloundeR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
