% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcell.R
\name{Flowcell}
\alias{Flowcell}
\title{R6 Class for performing Flowcell centric analyses}
\description{
This class aims to simplify the handling and exploration of Flowcell based
data and contains various presets, designs and visualisation tools required
for assessing flowcell performance and metrics.
}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{Flowcell}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Flowcell$new()}}
\item \href{#method-set_channel_counts}{\code{Flowcell$set_channel_counts()}}
\item \href{#method-clone}{\code{Flowcell$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print">}\href{../../floundeR/html/FloundeR.html#method-print}{\code{floundeR::FloundeR$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flowcell$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_channel_counts"></a>}}
\if{latex}{\out{\hypertarget{method-set_channel_counts}{}}}
\subsection{Method \code{set_channel_counts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flowcell$set_channel_counts(channel_counts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flowcell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
