% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infographic.R
\name{InfographicItem}
\alias{InfographicItem}
\title{R6 Class for loading and analysing sequence sets}
\description{
R6 Class for loading and analysing sequence sets

R6 Class for loading and analysing sequence sets
}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{InfographicItem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.key}}{the infographic key e.g. ReadCount}

\item{\code{.value}}{the element's value e.g. 42}

\item{\code{.icon}}{the fa-awesome code to use for the cartoon display}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InfographicItem-new}{\code{InfographicItem$new()}}
\item \href{#method-InfographicItem-clone}{\code{InfographicItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="bin_data"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-bin_data'><code>floundeR::FloundeR$bin_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="num_scale"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-num_scale'><code>floundeR::FloundeR$num_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="nums_scale"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-nums_scale'><code>floundeR::FloundeR$nums_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print"><a href='../../floundeR/html/FloundeR.html#method-FloundeR-print'><code>floundeR::FloundeR$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfographicItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-InfographicItem-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance of the R6 Class \code{InfographicItem}

This class is used to contain the information that is subsequently
rendered by the \code{Infographic} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfographicItem$new(key = NA, value = NA, icon = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{the infographic key e.g. ReadCount}

\item{\code{value}}{the element's value e.g. 42}

\item{\code{icon}}{the fa-awesome code to use for the cartoon display}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfographicItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InfographicItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfographicItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
