% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_summary.R
\name{SequencingSummary}
\alias{SequencingSummary}
\title{R6 Class for loading and analysing nanopore sequencing_summary files}
\description{
This class aims to simplify the handling and exploration of
sequencing_summary files and provides simple methods for accessing
information that can be used to assess the performance of a run.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sequencing_summary_file}}{the file.path
to the query FAST5 file}

\item{\code{barcoding_summary_file}}{the file.path
to the query FAST5 file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SequencingSummary$new()}}
\item \href{#method-clone}{\code{SequencingSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new SequencingSummary object. This
initialisation method performs other sanity checking
of the defined file(s) to ensure that it is indeed
parseable and creates the required data structures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$new(
  sequencing_summary_file = NA,
  barcoding_summary_file = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sequencing_summary_file}}{The source
sequencing_summary file.}

\item{\code{barcoding_summary_file}}{The source
barcoding_summary_file file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Fast5} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
