% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_summary.R
\name{SequencingSummary}
\alias{SequencingSummary}
\title{R6 Class for loading and analysing nanopore sequencing_summary files}
\description{
This class aims to simplify the handling and exploration of
sequencing_summary files and provides simple methods for accessing
information that can be used to assess the performance of a run.
}
\examples{

## ------------------------------------------------
## Method `SequencingSummary$new`
## ------------------------------------------------

sequencing_summary <- flnDr("sequencing_summary.txt.bz2")
barcodes_summary <- flnDr("barcoding_summary.txt.bz2")
seqsum <- SequencingSummary$new(sequencing_summary, barcodes_summary)
}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{SequencingSummary}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sequencing_summary_file}}{the file.path
to the query FAST5 file}

\item{\code{barcoding_summary_file}}{the file.path
to the query FAST5 file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SequencingSummary$new()}}
\item \href{#method-tibble}{\code{SequencingSummary$tibble()}}
\item \href{#method-clone}{\code{SequencingSummary$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print">}\href{../../floundeR/html/FloundeR.html#method-print}{\code{floundeR::FloundeR$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new SequencingSummary object. This
initialisation method performs other sanity checking
of the defined file(s) to ensure that it is indeed
parseable and creates the required data structures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$new(sequencing_summary_file, barcoding_summary_file = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sequencing_summary_file}}{The source
sequencing_summary file.}

\item{\code{barcoding_summary_file}}{The source
barcoding_summary_file file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Fast5} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{sequencing_summary <- flnDr("sequencing_summary.txt.bz2")
barcodes_summary <- flnDr("barcoding_summary.txt.bz2")
seqsum <- SequencingSummary$new(sequencing_summary, barcodes_summary)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tibble"></a>}}
\if{latex}{\out{\hypertarget{method-tibble}{}}}
\subsection{Method \code{tibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequencingSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
