% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_set.R
\name{TemporalSet}
\alias{TemporalSet}
\title{R6 Class for analysing sequence sets with accompanying temporal data}
\description{

}
\section{Super class}{
\code{\link[floundeR:FloundeR]{floundeR::FloundeR}} -> \code{TemporalSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TemporalSet$new()}}
\item \href{#method-as_tibble}{\code{TemporalSet$as_tibble()}}
\item \href{#method-runtime}{\code{TemporalSet$runtime()}}
\item \href{#method-run_yield}{\code{TemporalSet$run_yield()}}
\item \href{#method-feature_over_time}{\code{TemporalSet$feature_over_time()}}
\item \href{#method-t50}{\code{TemporalSet$t50()}}
\item \href{#method-clone}{\code{TemporalSet$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="bin_data">}\href{../../floundeR/html/FloundeR.html#method-bin_data}{\code{floundeR::FloundeR$bin_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="floundeR" data-topic="FloundeR" data-id="print">}\href{../../floundeR/html/FloundeR.html#method-print}{\code{floundeR::FloundeR$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalSet$new(seqsum = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalSet$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runtime"></a>}}
\if{latex}{\out{\hypertarget{method-runtime}{}}}
\subsection{Method \code{runtime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalSet$runtime()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_yield"></a>}}
\if{latex}{\out{\hypertarget{method-run_yield}{}}}
\subsection{Method \code{run_yield()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalSet$run_yield(resolution = 15, bases = TRUE, cumulative = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feature_over_time"></a>}}
\if{latex}{\out{\hypertarget{method-feature_over_time}{}}}
\subsection{Method \code{feature_over_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalSet$feature_over_time(
  resolution = 60,
  passes = TRUE,
  feature = "speed"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-t50"></a>}}
\if{latex}{\out{\hypertarget{method-t50}{}}}
\subsection{Method \code{t50()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalSet$t50(passes = TRUE, t = 0.5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
